refac: renaming folder and datadetailpage
