+++ test-web/src/components/DataDetailPage/InputTransitionsModal.js
@@ -3,18 +3,13 @@ import { Backdrop, Box, Modal, Fade, Button, Typography } from '@mui/material';
 
 const navy = '#0F3659';
 
-//엑셀 파일 업로드 실패/ 성공 여부 모달창
-const ExcelImportAlertModal = ({
-  isImportSuccessed,
-  setAlertDone,
-  setIsImportSuccessed,
-}) => {
+// 처리육 데이터 수정 전, 이미지 먼저 업로드 필수 알림 경고창
+const InputTransitionsModal = ({ setModal }) => {
   //화면 창 닫기
   const [open, setOpen] = useState(true);
   const handleClose = () => {
-    setAlertDone(false);
-    setIsImportSuccessed(true);
     setOpen(false);
+    setModal(false);
   };
 
   return (
@@ -35,37 +30,24 @@ const ExcelImportAlertModal = ({
         <Fade in={open}>
           <Box sx={style}>
             <Typography id="transition-modal-title" variant="h6" component="h2">
-              {isImportSuccessed ? (
-                <span
-                  style={{ color: navy, fontSize: '20px', fontWeight: '600' }}
-                >
-                  엑셀 파일 업로드에 성공하였습니다.
-                </span>
-              ) : (
-                <span
-                  style={{ color: navy, fontSize: '20px', fontWeight: '600' }}
-                >
-                  엑셀 파일 업로드에 실패하였습니다.
-                </span>
-              )}
+              <span
+                style={{ color: navy, fontSize: '20px', fontWeight: '600' }}
+              >
+                이미지를 먼저 업로드해야 합니다.
+              </span>
             </Typography>
-
             <div
-              style={{ display: 'flex', width: '100%', justifyContent: 'end' }}
+              style={{
+                display: 'flex',
+                width: '100%',
+                justifyContent: 'center',
+                marginTop: '20px',
+              }}
             >
               <Button
-                variant="outlined"
-                sx={{ marginRight: '5px' }}
+                variant="contained"
+                sx={{ marginRight: '5px', backgroundColor: navy }}
                 onClick={handleClose}
-                style={{
-                  backgroundColor: 'white',
-                  marginRight: '10px',
-                  border: `1px solid ${navy}`,
-                  height: '35px',
-                  borderRadius: '10px',
-                  width: '100px',
-                  fontSize: '17px',
-                }}
               >
                 닫기
               </Button>
@@ -75,7 +57,7 @@ const ExcelImportAlertModal = ({
       </Modal>
     </div>
   );
-};
+}
 
 const style = {
   position: 'absolute',
@@ -83,10 +65,11 @@ const style = {
   left: '50%',
   transform: 'translate(-50%, -50%)',
   width: 500,
+  height: '160px',
   bgcolor: 'background.paper',
   //border: '2px solid #000',
   boxShadow: 24,
   p: 4,
   borderRadius: '10px',
 };
-export default ExcelImportAlertModal
+export default InputTransitionsModal;
