+++ test-web/src/components/DataDetailPage/navBar.js
@@ -4,7 +4,7 @@ import { FaAngleDoubleLeft } from 'react-icons/fa';
 import ExcelController from '../DataListView/excelContr';
 
 // 상세 페이지에서 목록페이지로 이동 컴포넌트
-function navBar() {
+const NavBar = () => {
   return (
     <Box sx={{ display: 'flex', marginTop: '70px' }}>
       <div
@@ -28,6 +28,6 @@ function navBar() {
       </div>
     </Box>
   );
-}
+};
 
-export default navBar;
+export default NavBar;
