+++ test-web/src/components/DataDetailPage/restrictedModal.js
@@ -4,7 +4,7 @@ import { Backdrop, Box, Modal, Fade, Button, Typography } from '@mui/material';
 const navy = '#0F3659';
 
 // 이미지 수정 API 전송 실패 모달창
-export default function RestrictedModal({ setIsLimitedToChangeImage }) {
+const RestrictedModal = ({ setIsLimitedToChangeImage }) => {
   //화면 창 닫기
   const [open, setOpen] = useState(true);
   const handleClose = () => {
@@ -73,6 +73,7 @@ export default function RestrictedModal({ setIsLimitedToChangeImage }) {
     </div>
   );
 }
+export default RestrictedModal
 
 const style = {
   position: 'absolute',
